separate Component from  Article to  ArticleCard
