+++ pages/index.js
@@ -1,4 +1,4 @@
-import Article from "@/components/Article.js/Article";
+import Article from "@/components/Article.js/ArticleList";
 import Navigation from "@/components/Navigation/Navigation";
 
 const fetcher = (url) => fetch(url).then((res) => res.json());
